move sanitise helper funcs to the links package and adding IsValidInterfaceName
