+++ links/link.go
@@ -10,7 +10,6 @@ import (
 	"github.com/google/uuid"
 	"github.com/srl-labs/containerlab/internal/slices"
 	"github.com/srl-labs/containerlab/nodes/state"
-	"github.com/srl-labs/containerlab/utils"
 	"github.com/vishvananda/netlink"
 	"gopkg.in/yaml.v2"
 )
@@ -404,18 +403,19 @@ const (
 func SetNameMACAndUpInterface(l netlink.Link, endpt Endpoint) func(ns.NetNS) error {
 	return func(_ ns.NetNS) error {
 		// rename the link created with random name if its length is acceptable by linux
-		if len(endpt.GetIfaceName()) < 16 {
+		if IsValidInterfaceName(endpt.GetIfaceName()) {
 			err := netlink.LinkSetName(l, endpt.GetIfaceName())
 			if err != nil {
 				return fmt.Errorf(
 					"failed to rename link: %v", err)
 			}
 		} else {
-			// when the name is too long, we add an AltName instead of a regular interface name
-			err := netlink.LinkAddAltName(l, endpt.GetIfaceName())
+			// when the name is too long, we add a sanitized interface name as AltName
+			sanitisedIfaceName := SanitiseInterfaceName(endpt.GetIfaceAlias())
+			err := netlink.LinkAddAltName(l, sanitisedIfaceName)
 			if err != nil {
 				return fmt.Errorf(
-					"failed to add alias: %v", err)
+					"failed to add altname: %v", err)
 			}
 		}
 
@@ -433,7 +433,7 @@ func SetNameMACAndUpInterface(l netlink.Link, endpt Endpoint) func(ns.NetNS) err
 				return err
 			}
 			// Set a sanitised altname for ease of access. '/', and ' ' are changed to '-'
-			sanitisedIfaceName := utils.SanitiseInterfaceName(endpt.GetIfaceAlias())
+			sanitisedIfaceName := SanitiseInterfaceName(endpt.GetIfaceAlias())
 			err = netlink.LinkAddAltName(l, sanitisedIfaceName)
 			if err != nil {
 				return err
@@ -490,6 +490,39 @@ func isInFilter(params *ResolveParams, endpoints []*EndpointRaw) bool {
 		if !slices.Contains(params.NodesFilter, e.Node) {
 			return false
 		}
+	}
+
+	return true
+}
+
+// SanitiseInterfaceName sanitises the interface name by replacing '/' and ' ' with '-'.
+// Making it suitable to write as AltName for the interface.
+func SanitiseInterfaceName(ifaceName string) string {
+	var sb strings.Builder
+	sb.Grow(len(ifaceName))
+
+	for _, char := range ifaceName {
+		switch char {
+		case '/', ' ':
+			sb.WriteRune('-')
+		default:
+			sb.WriteRune(char)
+		}
+	}
+
+	return sb.String()
+}
+
+// IsValidInterfaceName checks if the interface name is valid
+// by checking its length and the presence of spaces or slashes.
+func IsValidInterfaceName(ifaceName string) bool {
+	if len(ifaceName) > 15 {
+		return false
+	}
+
+	// interface name can not contain spaces or slashes
+	if strings.ContainsAny(ifaceName, " /") {
+		return false
 	}
 
 	return true
